完善系统功能，后续功能抽时间再加

+++ src/main/java/com/spring/controller/permission/PMenuController.java
@@ -1,10 +1,14 @@
 package com.spring.controller.permission;
 
 import com.github.pagehelper.Page;
+import com.spring.common.exceptions.MyException;
 import com.spring.common.utils.ResponseUtils;
 import com.spring.controller.base.SuperController;
 import com.spring.model.permission.Menu;
+import com.spring.model.permission.RoleMenu;
+import com.spring.param.MenuFilter;
 import com.spring.service.permission.MenuService;
+import com.spring.service.permission.RoleMenuService;
 import org.apache.ibatis.session.RowBounds;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.slf4j.Logger;
@@ -12,11 +16,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
+import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -34,31 +36,34 @@ public class PMenuController extends SuperController {
     @Autowired
     private MenuService menuService;
 
+    @Autowired
+    private RoleMenuService roleMenuService;
+
     //加载page页面
     @RequiresPermissions(value = "permission:menu:page")
-    @RequestMapping(value = "/page",method = RequestMethod.GET)
-    public String page(){
-        return "permission/menu/page";
-    }
-
-    //加载page页面中的date数据
-    @RequestMapping(value = "/data",method = RequestMethod.POST)
-    public String date(Integer pageNo, @RequestParam(defaultValue = "10")Integer pageSize, Model model){
+    @RequestMapping(value = "/page")
+    public String page(@RequestParam(defaultValue = "1")Integer pageNo,
+                       @RequestParam(defaultValue = "10")Integer pageSize,
+                       Model model){
+        //获取所有的菜单名称
+        List<Menu> menuList=menuService.getAllMenus();
+        model.addAttribute("menuNameList",menuList);
+        //获取菜单数据
         Page<Menu> lists=menuService.getAllMenusPage(new RowBounds((pageNo-1)*pageSize,pageSize));
         model.addAttribute("lists",lists);
-        return "permission/menu/data";
+        return "permission/menu/menuList";
     }
 
     //加载菜单编辑页面
-    @RequestMapping(value = "toUpdate",method = RequestMethod.GET)
+    @RequestMapping(value = "/toUpdate",method = RequestMethod.GET)
     public String toAdd(Model model){
         List<Menu> lists=menuService.getAllFirstMenus();
         model.addAttribute("menus",lists);
-        return "permission/menu/update";
+        return "permission/menu/menu-add";
     }
 
     //加载菜单修改页面
-    @RequestMapping(value = "{menuId}/toUpdate",method = RequestMethod.GET)
+    @RequestMapping(value = "/{menuId}/toUpdate",method = RequestMethod.GET)
     public String toUpdate(@PathVariable("menuId")String menuId, Model model){
         Menu menu=menuService.getMenuByMenuId(menuId);
         if(menu.getParentId()!=null&&!"".equals(menu.getParentId())){
@@ -69,11 +74,11 @@ public class PMenuController extends SuperController {
         List<Menu> list=menuService.getAllFirstMenus();
         model.addAttribute("menus",list);
         model.addAttribute("menu",menu);
-        return "permission/menu/update";
+        return "permission/menu/menu-add";
     }
 
     //对菜单信息数据更改
-    @RequestMapping(value = "update",method = RequestMethod.POST)
+    @RequestMapping(value = "/update",method = RequestMethod.POST)
     public void update(Menu menu){
         if(menu.getMenuId()==null){
             //新增操作
@@ -85,10 +90,49 @@ public class PMenuController extends SuperController {
         }
     }
 
-    @RequestMapping(value = "{menuId}/delete",method = RequestMethod.POST)
-    public void delete(@PathVariable("menuId") String menuId){
+    @RequestMapping(value = "/delete",method = RequestMethod.POST)
+    public void delete(String menuId){
+        List<RoleMenu> list=roleMenuService.getMessageByMenuId(menuId);
+        if (list!=null&&list.size()>0){
+            ResponseUtils.writeErrorResponse(request,response,"存在角色菜单关联信息，无法删除");
+            throw new MyException("存在角色菜单关联信息，无法删除");
+        }
         menuService.delete(menuId);
         ResponseUtils.writeSuccessReponse(request,response,"删除菜单信息成功");
     }
+
+    @PostMapping(value = "/batchDelete")
+    public void batchDelete(String checkedId){
+        String[] record=checkedId.split(",");
+        List<String> checkList=Arrays.asList(record);
+        for (String menuId: checkList){
+            List<RoleMenu> list=roleMenuService.getMessageByMenuId(menuId);
+            if (list!=null&&list.size()>1){
+                ResponseUtils.writeErrorResponse(request,response,"存在角色菜单关联信息，无法删除");
+                throw new MyException("存在角色菜单关联信息，无法删除");
+            }
+        }
+        menuService.batchDelete(checkList);
+        ResponseUtils.writeSuccessReponse(request,response,"批量删除菜单信息成功");
+    }
+
+    @PostMapping(value = "/start")
+    public void start(String menuId){
+        Menu menu=menuService.getMenuByMenuId(menuId);
+        menu.setStatus(menu.getStatus()^1);
+        menuService.update(menu);
+        ResponseUtils.writeSuccessReponse(request,response,"成功啦");
+    }
+
+    @PostMapping(value = "/retrieve")
+    public String retrieve(MenuFilter filter,Model model){
+        Integer pageNo=filter.getPageNo();
+        Integer pageSize=filter.getPageSize();
+        Page<Menu> lists=menuService.getMessageByCondition(new RowBounds((pageNo-1)*pageSize,pageSize),filter);
+        model.addAttribute("lists",lists);
+        return "permission/menu/menuList";
+
+    }
+
 
 }
